Try to minimize the external dependencies for OMPI wrappers.
The solution put forward here consist of creating a convenience library,
opal-tiny, that holds all the basic things we need to have a working
app with minimal external dependencies.

More detailed discussion in #9869.
